requirements.txt
flask
langchain-openai
python-dotenv
markdown